factor our readGmshFileV2 into a separate file
